+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Remove the default timeout of 60 seconds from the Context
 * Add `bool` return type to `fingerprint()` function to indicate if the callable was run
 * Add a `recursive` parameter to the `withData()` method of `Context` to allow recursive merging for nested arrays
+* Add an `open()` function to open a file or URL in the default application
 
 ## 0.13.1 (2024-02-27)